Move cert-manager.io/cluster-issuer to deployments.
We need to set this for all hubs except cdss-discovery. I used to unset the value in cdss-discovery by setting it to null, but now z2jh emits a validation error, complainting that it isn't a string. There's no way to unset a value -- you can only set it to something that validates. In this case, setting it to an empty string wouldn't be correct. So we have to set it for all hubs except cdss-discovery.